+++ bin/streamr-stream-list.ts
@@ -8,6 +8,7 @@ const program = new Command();
 program
     .description('fetch a list of streams that are accessible by the authenticated user')
     .option('-s, --search [term]', 'search for term in name or description')
+    // TODO could support shorter forms of operations: e.g. "publish" instead of "stream_publish", see streamr-stream-grant-permission.ts
     .addOption(new Option('-o, --operation [permission]', 'filter by permission').choices(['stream_get','stream_subscribe','stream_publish','stream_delete','stream_share']).default('stream_get'))
     .option('--public-access', 'include publicly available streams')
     .option('--no-granted-access', 'exclude streams that user has directly granted permissions to')