+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Add permission commands: `stream grant-permission` and `stream revoke-permission`
 
 ## [5.0.0] - 2021-05-05
 ### Added